Committed after changes from int to intptr in user32.dll class
git-svn-id: http://www.aarp-tax-aide-sw-washington.org/svnrepo/scripts-ng/trunk/Tax-Aide TrueCrypt Utility@194 b9d85104-183b-4f64-83c0-99b67ab08840

+++ Tax-Aide TrueCrypt Utility/TcAction.cs
@@ -125,7 +125,7 @@ namespace TaxAide_TrueCrypt_Utility
             {   //a fail due to non empty directory so can do install
                 Win32.PostMessage(tcWin.hWndHandle, (int)win32Message.WM_KEYDOWN, (IntPtr)'\r', (IntPtr)0x1C0001); //sends enter to dialog box
                 Thread.Sleep(1000);
-                int hWnd = Win32.FindWindow("CustomDlg", TCWin.mainWinTitle); //now send escape to mainW window
+                IntPtr hWnd = Win32.FindWindow("CustomDlg", TCWin.mainWinTitle); //now send escape to mainW window
                 Win32.PostMessage(hWnd, (int)win32Message.WM_KEYDOWN, (IntPtr)'\u001B', (IntPtr)0x10001); //sends enter to dialog box
                 TCWin.uninstallFailContinue = true;
             }
